Updated documetation and licensing  - added new supported models section  - migrated open-mmlab, and dreambooth configs to corresponding directories
