+++ docs/api/hub.md
@@ -5,5 +5,3 @@
 ## ::: nos.hub.config.TorchHubConfig
 
 ## ::: nos.hub.config.HuggingFaceHubConfig
-
-## ::: nos.hub.config.MMLabConfig
