+++ mkdocs.yml
@@ -15,7 +15,7 @@ site_dir: site
 nav:
   - 🏠 Home: README.md
   - 🔥 Quickstart: docs/quickstart.md
-  - 🧠 Models: docs/models.md
+  - 🧠 Models: docs/models/supported-models.md
   - ⚡️ Concepts:
     - Why NOS?: docs/concepts/what-is-nos.md
     - Architecture: docs/concepts/architecture-overview.md
@@ -41,8 +41,8 @@ nav:
     - nos.executors: docs/api/executors.md
   - 💻 CLI Reference:
     - nos system: docs/cli/system.md
-  #   - nos hub: docs/cli/hub.md
-  #   - nos predict: docs/cli/predict.md
+    - nos hub: docs/cli/hub.md
+    - nos predict: docs/cli/predict.md
   #   - nos benchmark: docs/cli/benchmark.md
   #   - nos docker: docs/cli/docker.md
   - 🔌 Integrations:
