+++ nos/models/__init__.py
@@ -11,10 +11,8 @@ from nos.common.types import Batch, ImageT
 from ._noop import NoOp  # noqa: F401
 from .blip import BLIP  # noqa: F401
 from .clip import CLIP  # noqa: F401
-from .dreambooth.dreambooth import StableDiffusionLoRA  # noqa: F401
 from .faster_rcnn import FasterRCNN  # noqa: F401
 from .monodepth import MonoDepth  # noqa: F401
-from .openmmlab.mmdetection import MMDetection  # noqa: F401
 from .owlvit import OwlViT  # noqa: F401
 from .sam import SAM
 from .stable_diffusion import StableDiffusion  # noqa: F401
