--- nos/server/train/__init__.py
@@ -1,1 +1,0 @@
-from ._service import TrainingService  # noqa: F401
